pyproject.toml
[project]
name = "crypto-portfolio-tracker"
authors = [
  { name = "crypto-portfolio-tracker", email = "king.arthur@camelot.bt" }
]
description = "Fetch DeFi staking/lending positions for any wallet address across multiple chains using pure RPC calls"
readme = "README.md"
dynamic = ["version"]
license = { file = "LICENSE" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: Implementation :: CPython",
]
requires-python = ">=3.11"
dependencies = [
    "eth-ape",
    "ape-infura",
    "ape-etherscan",
    "ape-base",
    "typer",
    "rich-click",
    "pydantic",
    "pyyaml",
    "httpx",
]

[project.scripts]
crypto-portfolio-tracker = "crypto_portfolio_tracker.cli.main:app"

[project.urls]
Documentation = "https://king-arthur.github.io/crypto-portfolio-tracker/"
Source = "https://github.com/king-arthur/crypto-portfolio-tracker"

[tool.hatch]

[tool.hatch.metadata]
allow-direct-references = true


##############
# Versioning #
##############

[tool.hatch.version]
source = "regex_commit"
commit_extra_args = ["-e"]
path = "scripts/__init__.py"


##################
# External Tools #
##################

[tool.mypy]
files = ["crypto_portfolio_tracker"]
exclude = ["build/", "dist/", "docs/", "tests/*"]
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true


[tool.ruff]
target-version = "py311"
line-length = 120
indent-width = 4
include = [
  "src/**/*.py",
  "src/**/*.pyi",
  "tests/**/*.py",
  "tests/**/*.pyi"
]
exclude = ["tests"]

[tool.ruff.lint]
preview = true # preview features & checks, use with caution
extend-select = [ # features in preview
    "W292", # missing-newline-at-end-of-file
]
select = [
    "A",
    "ARG",
    "B",
    "C",
    "DTZ",
    "E",
    "EM",
    "F",
    "FBT", # Boolean trap
    "ICN",
    "ISC",
    "I",
    "N",
    "PLC",
    "PLE",
    "PLR",
    "PLW",
    "Q",
    "RUF",
    "S",
    "T",
    "TID",
    "UP",
    "W",
    "YTT",
    "RUF100", # Automatically remove unused # noqa directives
]
ignore = [
    # Allow non-abstract empty methods in abstract base classes
    "B027",
    # Allow boolean positional values in function calls, like `dict.get(... True)`
    "FBT003",
    # Ignore checks for possible passwords
    "S105", "S106", "S107",
    # Ignore complexity
    "C901", "PLR0911", "PLR0912", "PLR0913", "PLR0915",
    "PLC1901", # empty string comparisons
    "PLW2901", # `for` loop variable overwritten
    "SIM114", # Combine `if` branches using logical `or` operator
    "E203", # Whitespace before :, needed for black compatability and also `ruff format`
    "ISC001", # causes unexpected behaviour with formatter
]


[tool.ruff.lint.isort]
known-first-party = ["crypto_portfolio_tracker"]


[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"


[tool.ruff.lint.per-file-ignores]
# Allow print/pprint
"examples/*" = ["T201"]
# Tests can use magic values, assertions, and relative imports
"tests/**/*" = ["PLR2004", "S101", "TID252"]


[tool.pytest.ini_options]
# addopts = "--cov=src/crypto_portfolio_tracker/ --cov-report=term-missing"
# Disable ape network connection during tests
addopts = "-p no:ape"


#####################
# Environment Setup #
#####################


# Default env dendencies
[tool.hatch.envs.default]
extra-dependencies = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]
installer = "uv"
# post-install-commands = ["pre-commit install"]


# Lint env dendencies
[tool.hatch.envs.lint]
extra-dependencies = [
  "mypy",
  "ruff",
]


[tool.hatch.envs.lint.scripts]
typing = [
  "echo \"VERSION: `mypy --version`\"",
  "mypy --install-types --non-interactive {args}"
]
lint = [
  "echo \"VERSION: `ruff --version`\"",
  "ruff format .",
  "ruff check . --fix",
  "mypy src/crypto_portfolio_tracker/",
]
lint-check = [
  "echo \"VERSION: `ruff --version`\"",
  "ruff format --check .",
  "ruff check .",
  "mypy src/crypto_portfolio_tracker/",
]

# Docs env dendencies
[tool.hatch.envs.docs]
dependencies = [
"mkdocs",
  "mkdocs-material",
  "mkdocs-gen-files",
  "mkdocstrings[python]",
  "linkchecker",
]


[tool.hatch.envs.docs.scripts]
docs-serve = "mkdocs serve"
docs-build = "mkdocs build --clean --strict"
# --ignore-url=None since the SUMMARY.md file leaves a <toc>None</toc> in sitemap.xml
validate = "linkchecker --config .linkcheckerrc --ignore-url=/reference --ignore-url=None site"
# https://github.com/linkchecker/linkchecker/issues/678
build-check = [
  "docs-build",
  "validate",
]

# Testing env dendencies
[tool.hatch.envs.test]
extra-dependencies = [
  "pytest",
]


[tool.hatch.envs.test.scripts]
test = "pytest"
test-cov-xml = "pytest --cov-report=xml"


# Replace tox & nox
[[tool.hatch.envs.test.matrix]]
python = ["3.11", "3.12"]

#########
# Build #
#########

[build-system]
requires = ["hatchling", "hatch-regex-commit"]
build-backend = "hatchling.build"


